typecasts to please picky compilers checking the printf() format string

+++ src/urlglob.c
@@ -18,7 +18,7 @@
  * This software is distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY
  * KIND, either express or implied.
  *
- * $Id: urlglob.c,v 1.32 2004-03-23 09:12:51 bagder Exp $
+ * $Id: urlglob.c,v 1.33 2004-04-06 07:48:29 bagder Exp $
  ***************************************************************************/
 
 /* client-local setup.h */
@@ -77,13 +77,13 @@ static GlobCode glob_set(URLGlob *glob, char *pattern,
     switch (*pattern) {
     case '\0':			/* URL ended while set was still open */
       snprintf(glob->errormsg, sizeof(glob->errormsg),
-               "unmatched brace at pos %d\n", pos);
+               "unmatched brace at pos %d\n", (int)pos);
       return GLOB_ERROR;
 
     case '{':
     case '[':			/* no nested expressions at this time */
       snprintf(glob->errormsg, sizeof(glob->errormsg),
-               "nested braces not supported at pos %d\n", pos);
+               "nested braces not supported at pos %d\n", (int)pos);
       return GLOB_ERROR;
 
     case ',':
@@ -119,13 +119,13 @@ static GlobCode glob_set(URLGlob *glob, char *pattern,
 
     case ']':				/* illegal closing bracket */
       snprintf(glob->errormsg, sizeof(glob->errormsg), 
-               "illegal pattern at pos %d\n", pos);
+               "illegal pattern at pos %d\n", (int)pos);
       return GLOB_ERROR;
 
     case '\\':				/* escaped character, skip '\' */
       if (*(buf+1) == '\0') {		/* but no escaping of '\0'! */
         snprintf(glob->errormsg, sizeof(glob->errormsg), 
-                 "illegal pattern at pos %d\n", pos);
+                 "illegal pattern at pos %d\n", (int)pos);
 	return GLOB_ERROR;
       }
       ++pattern;
